voice-agent-workshop/livekit-voice-agent/pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-openai~=1.3",
    "livekit-plugins-assemblyai~=1.3",
    "livekit-plugins-cartesia~=1.3",
    "python-dotenv>=1.1.1",
    "aiohttp>=3.9.0",
]
